Extract registers
